add mars.tensor.statistics, move average, corrcoef, cov, ptp, histogram_bin_edges, histogram into it
